Keep `z.optional` and `z.discriminatedUnion` compatible with covalues
